refactor: split ResultCard.tsx into sub-components
Extract ResultCard (741 lines, 34 functions) into focused sub-components:
- ResultCardHeader: thumbnail, badges, name, share/expand
- ResultCardRelease: featured release with preview + links
- ResultCardPlatforms: categorized platform badges
- ResultCardPreview: Bandcamp embed player
- ResultCardSocial: social icon links
- ResultCardActions: app promo, claim, report, legend
- ResultCardUtils: shared helpers (getSource, isDirectLink, categorizePlatforms, getReleaseInfo)
- ResultCardTypes: shared interfaces

All existing behavior preserved. Analytics tracking intact.
